finalizado primer test

+++ src/Interprete/decoder.py
@@ -20,7 +20,7 @@ class deco:
     def getDataDegree(self,degree):
         if degree <= 360:
             unity_conv = 0.01
-            pos_LSB = degree/unity_conv #convercion a rago del motor (LSB = low significant bit)
+            pos_LSB = int(degree/unity_conv) #convercion a rago del motor (LSB = low significant bit)
             #data to motor
             data_degree = bytearray([
                 pos_LSB & 0xFF,
